pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "ucl-open"
dynamic = ["version"]
description = "Environment to generate and extend shared and reuseable schemas and a DSL for ucl-open experiments"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.12.3",
    "pydantic-yaml>=1.6.0",
    "swc-aeon>=0.2.0",
]

[project.optional-dependencies]
dev = [
  "pandas-stubs",
  "pre-commit",
  "pyright",
  "pytest",
  "pytest-cov",
  "pytest-mock",
  "ruff",
]

[tool.ruff]
line-length = 108

[tool.pyright]
useLibraryCodeForTypes = false
reportMissingImports = "error"
reportIncompatibleVariableOverride = "error"
reportImportCycles = "error"
reportUnusedImport = "error"
reportUnusedClass = "error"
reportUnusedFunction = "error"
reportUnusedVariable = "error"
reportDuplicateImport = "error"
reportWildcardImportFromLibrary = "error"
reportPrivateUsage = "error"
reportCallInDefaultInitializer = "error"
reportUnnecessaryIsInstance = "error"
reportUnnecessaryCast = "error"
reportUnnecessaryComparison = "error"
reportUnnecessaryContains = "error"
reportAssertAlwaysTrue = "error"
reportSelfClsParameterName = "error"
reportUnusedExpression = "error"
reportMatchNotExhaustive = "error"
# *Note*: we may want to set all 'ReportOptional*' rules to "none", but leaving 'em default for now
venvPath = "."
venv = ".venv"
exclude = [".venv/*"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = [
  "--cov=src/ucl_open",    # Enable pytest-cov for the ucl_open package
]
